test: move pydantic tests from test_data_parsing
